backend/requirements.txt
# --- Backend Frameworks ---
fastapi
flask

# ASGI server for FastAPI
uvicorn[standard]

# Environment variables
python-dotenv

# Database
sqlalchemy
alembic
psycopg2-binary

# Validation
pydantic

# File uploads
python-multipart

# Document processing
python-docx
pymupdf
numpy

# AI / NLP
torch
transformers
accelerate
requests

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart